requirements.txt
PyQt6>=6.6,<7
pywinauto>=0.6.8,<1
sentence-transformers>=3,<4
PyYAML>=6,<7
spacy>=3.7,<4
chromadb>=0.5,<1
